moved f-n addDish
