+++ src/components/card/card.js
@@ -1,5 +1,7 @@
 import { getMenu } from "../../api/getApi";
 import { addMenu } from "./createCards";
+import { getCardDatabById } from "../../api/getApi";
+import { setOrdersDataToStorage, getOrdersDataToStorage } from "../../utils/localStorage";
 
 export async function getCardsData(value) {
   try {
@@ -38,6 +40,29 @@ function selectedCategory(event) {
   deleteStylesBtnsCategories();
   btnCategory.classList.add("menu__item-button_active");
   getCardsData(dataCategory);
+}
+
+function addDish(event) {
+  if(event.target.classList.contains('card__btn')) {
+      const cardId = +event.currentTarget.id;
+
+      try {
+          getCardDatabById(cardId)
+          .then(dataOrder => {
+              const dataCard = getOrdersDataToStorage('Orders');
+
+              if(dataCard === null) {
+                  setOrdersDataToStorage([dataOrder]);
+              } else {
+                  const dataFromLocalStorage = getOrdersDataToStorage("Orders");
+                  dataFromLocalStorage.push(dataOrder);
+                  setOrdersDataToStorage(dataFromLocalStorage);
+              }
+          })
+      } catch(error) {
+        throw error;
+      }
+  }
 }
 
 document.addEventListener("DOMContentLoaded", () => {
